scripts/pyproject.toml
[tool.poetry]
name = "resolution-sort"
version = "0.1.0"
description = "A tool to sort wallpapers into their respective resolution categories"
authors = ["Marcus <psminchia4153@hotmail.com>"]
# readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^10.0.0"

[tool.poetry.scripts]
sort-walls = "resolution_sort:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
